SWARM-595 -- General cleanup of org.wildfly.swarm.runtime.**
Attempt to organize classes into logical packages.

Also remove a class or two that is no longer actually used.
