--- container/src/main/java/org/wildfly/swarm/container/runtime/DefaultInterfaceProducer.java
@@ -1,35 +1,0 @@
-/**
- * Copyright 2015-2016 Red Hat, Inc, and individual contributors.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- * http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package org.wildfly.swarm.container.runtime;
-
-import javax.enterprise.context.ApplicationScoped;
-import javax.enterprise.inject.Produces;
-
-import org.wildfly.swarm.container.Interface;
-import org.wildfly.swarm.spi.api.SwarmProperties;
-
-/**
- * @author Bob McWhirter
- */
-@ApplicationScoped
-public class DefaultInterfaceProducer {
-
-    @Produces
-    public Interface publicInterace() {
-        return new Interface( "public", SwarmProperties.propertyVar(SwarmProperties.BIND_ADDRESS, "0.0.0.0"));
-
-    }
-}
