+++ container/src/main/java/org/wildfly/swarm/container/runtime/RuntimeServer.java
@@ -70,8 +70,10 @@ import org.wildfly.swarm.container.Interface;
 import org.wildfly.swarm.container.internal.Deployer;
 import org.wildfly.swarm.container.internal.Server;
 import org.wildfly.swarm.container.runtime.deployments.DefaultDeploymentCreator;
-import org.wildfly.swarm.container.runtime.internal.marshal.DMRMarshaller;
-import org.wildfly.swarm.container.runtime.internal.xmlconfig.StandaloneXMLParser;
+import org.wildfly.swarm.container.runtime.marshal.DMRMarshaller;
+import org.wildfly.swarm.container.runtime.wildfly.SimpleContentProvider;
+import org.wildfly.swarm.container.runtime.wildfly.UUIDFactory;
+import org.wildfly.swarm.container.runtime.xmlconfig.StandaloneXMLParser;
 import org.wildfly.swarm.internal.FileSystemLayout;
 import org.wildfly.swarm.spi.api.ArchiveMetadataProcessor;
 import org.wildfly.swarm.spi.api.ArchivePreparer;
