+++ container/src/main/java/org/wildfly/swarm/container/runtime/marshal/ProjectStagePropertyMarshaller.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.container.runtime.internal.marshal;
+package org.wildfly.swarm.container.runtime.marshal;
 
 import java.util.List;
 import java.util.Map;
