+++ container/src/main/java/org/wildfly/swarm/container/runtime/wildfly/UUIDFactory.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.container.runtime;
+package org.wildfly.swarm.container.runtime.wildfly;
 
 import java.util.UUID;
