+++ container/src/main/java/org/wildfly/swarm/container/runtime/xmlconfig/StandaloneXMLParser.java
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.wildfly.swarm.container.runtime.internal.xmlconfig;
+package org.wildfly.swarm.container.runtime.xmlconfig;
 
 import java.io.IOException;
 import java.io.InputStream;
