+++ container/src/main/java/org/wildfly/swarm/container/runtime/xmlconfig/StandaloneXMLParserProducer.java
@@ -1,4 +1,4 @@
-package org.wildfly.swarm.container.runtime.internal.xmlconfig;
+package org.wildfly.swarm.container.runtime.xmlconfig;
 
 import java.util.ArrayList;
 import java.util.Iterator;
