+++ mod_cluster/src/main/java/org/wildfly/swarm/mod_cluster/runtime/ModclusterSocketBindingCustomizer.java
@@ -4,7 +4,7 @@ import javax.inject.Inject;
 import javax.inject.Named;
 import javax.inject.Singleton;
 
-import org.wildfly.swarm.container.runtime.DefaultSocketBindingGroupProducer;
+import org.wildfly.swarm.container.runtime.config.DefaultSocketBindingGroupProducer;
 import org.wildfly.swarm.mod_cluster.ModclusterProperties;
 import org.wildfly.swarm.spi.api.Customizer;
 import org.wildfly.swarm.spi.api.SocketBinding;
